DDF-4222 extracted subclasses
